functions, bash script, installer
Seperated functions into a seperate file, added init file so searchit is viewed as a package, added a setup file in case someone wants to manually build the project. Soon I'll be uploading a "installer" which  can  be used with pip. It will automatically install all dependencies and searchit! How awesome is that :D 
Some small notes are that all the functions can now be used in other projects if needed but pages can still be searched using the bash script or with the searchit python file the same way as before.