Special Characters Search
Fixed an issue where special characters like "&" were breaking the search because of special meaning in the url. This now works properly by converting them to their hex value but make sure you enter your search inside "" marks because some of them are also used in bash for special commands. This won't be an issue after a gui update which I plan on making soon.